TBC20: merge file TBC20 into file token as it is the helper class for token
